+++ ergo_script/src/testing.rs
@@ -24,6 +24,7 @@ impl Test {
                 .threads(Some(1))
                 .keep_going(false),
             vec![],
+            None,
         )
         .expect("failed to create runtime");